Split out colour combinations
