+++ .storybook/main.js
@@ -65,6 +65,6 @@ module.exports = {
     return config;
   },
   docs: {
-    autodocs: true,
+    autodocs: false,
   },
 };
