--- src/nationalarchives/stories/utilities/colour-schemes/colour-schemes.mdx
@@ -1,107 +1,0 @@
-import { Meta, Canvas } from "@storybook/blocks";
-import * as ColourStories from './colour-schemes.stories';
-
-<Meta of={ColourStories} />
-
-# Colour schemes
-
-## System theme
-
-The system theme follows the operating system preference for `prefers-color-scheme` which can either be `light` or `dark`.
-
-The system theme also follows the preference for `prefers-contrast` which can be `no-preference`, `more`, `less` or `custom`.
-
-At this time, only `prefers-contrast: no-preference` and `prefers-contrast: more` are supported.
-
-- [MDN Web Docs - prefers-color-scheme](https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-color-scheme)
-- [MDN Web Docs - prefers-contrast](https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-contrast)
-
-```html
-<html class="tna-template tna-template--system-theme">
-  <head>
-    <title>System theme</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-## Light theme
-
-```html
-<html class="tna-template tna-template--light-theme">
-  <head>
-    <title>Light theme</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-## Dark theme
-
-```html
-<html class="tna-template tna-template--dark-theme">
-  <head>
-    <title>Dark theme</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-## High contrast themes
-
-The high contrast themes are applied on top of the existing light and dark themes with an extra `tna-template--high-contrast-theme` class on the `tna-template` element.
-
-### Light, high contrast theme
-
-```html
-<html class="tna-template tna-template--light-theme tna-template--high-contrast-theme">
-  <head>
-    <title>Light, high contrast theme</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-### Dark, high contrast theme
-
-```html
-<html class="tna-template tna-template--dark-theme tna-template--high-contrast-theme">
-  <head>
-    <title>Dark, high contrast theme</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-## Colour definitions
-
-The colours are defined in `src/nationalarchives/variables/_colour.scss` as a separate map for each theme.
-
-This allows you to reference a colour, such as `font-base` without having to worry about which theme is currently in use.
-
-## Accent colours
-
-To add accent colours to the page, add one of the following classes to the `<html>` element:
-
-- `tna-template--yellow-accent`
-- `tna-template--pink-accent`
-- `tna-template--orange-accent`
-- `tna-template--green-accent`
-- `tna-template--blue-accent`
-
-```html
-<html class="tna-template tna-template--system-theme tna-template--yellow-accent">
-  <head>
-    <title>System theme with yellow accents</title>
-  </head>
-  <body class="tna-template__body">
-  </body>
-</html>
-```
-
-
