--- src/nationalarchives/stories/utilities/lists/lists.mdx
@@ -1,18 +1,0 @@
-import { Meta, Canvas } from "@storybook/blocks";
-import * as ListStories from './lists.stories';
-
-<Meta of={ListStories} />
-
-# Lists
-
-<Canvas of={ListStories.UnorderedList} />
-<Canvas of={ListStories.OrderedList} />
-<Canvas of={ListStories.DescriptionList} />
-<Canvas of={ListStories.DescriptionListWithIcons} />
-
-## Plain versions
-
-<Canvas of={ListStories.UnorderedListPlain} />
-<Canvas of={ListStories.OrderedListPlain} />
-<Canvas of={ListStories.PlainDescriptionList} />
-<Canvas of={ListStories.PlainDescriptionListWithIcons} />
