--- src/nationalarchives/stories/utilities/tables/tables.mdx
@@ -1,8 +1,0 @@
-import { Meta, Canvas } from "@storybook/blocks";
-import * as TableStories from './tables.stories';
-
-<Meta of={TableStories} />
-
-# Tables
-
-<Canvas of={TableStories.Table} />
