--- src/nationalarchives/stories/utilities/typography/typography.mdx
@@ -1,27 +1,0 @@
-import { Meta, Canvas } from "@storybook/blocks";
-import * as HeadingStories from './headings.stories';
-import * as HeadingGroupStories from './heading-groups.stories';
-import * as TypographyStories from './typography.stories';
-
-<Meta of={TypographyStories} />
-
-# Typography
-
-## Headings
-
-<Canvas of={HeadingStories.Headings} />
-<Canvas of={TypographyStories.HeadingLink} />
-
-## Heading groups
-
-<Canvas of={HeadingGroupStories.HeadingGroup} />
-<Canvas of={HeadingGroupStories.HeadingGroupSingleSentence} />
-<Canvas of={HeadingGroupStories.HeadingGroupPlainSupertitle} />
-
-## General typography
-
-<Canvas of={TypographyStories.Paragraph} />
-<Canvas of={TypographyStories.LargeParagraph} />
-<Canvas of={TypographyStories.SceneSetter} />
-<Canvas of={TypographyStories.SceneSetterSmall} />
-<Canvas of={TypographyStories.Blockquote} />
